+++ src/account/interface.ts
@@ -15,10 +15,12 @@ import {
   EstimateFeeAction,
   EstimateFeeDetails,
   EstimateFeeResponse,
+  EstimateFeeResponseBulk,
   InvocationsDetails,
   InvokeFunctionResponse,
   MultiDeployContractResponse,
   Signature,
+  TransactionBulk,
   UniversalDeployerContractPayload,
 } from '../types';
 import { BigNumberish } from '../utils/number';
@@ -77,10 +79,12 @@ export abstract class AccountInterface extends ProviderInterface {
   /**
    * Estimate Fee for executing a DEPLOY_ACCOUNT transaction on starknet
    *
-   * @param contractPayload the payload object containing:
+   * @param contractPayload -
    * - contract - the compiled contract to be deployed
    * - classHash - the class hash of the compiled contract. This can be obtained by using starknet-cli.
-   *
+   * @param estimateFeeDetails -
+   * - optional blockIdentifier
+   * - constant nonce = 0
    * @returns response from estimate_fee
    */
   public abstract estimateAccountDeployFee(
@@ -107,6 +111,21 @@ export abstract class AccountInterface extends ProviderInterface {
     deployContractPayload: UniversalDeployerContractPayload | UniversalDeployerContractPayload[],
     transactionsDetail?: InvocationsDetails
   ): Promise<EstimateFeeResponse>;
+
+  /**
+   * Estimate Fee for executing a list of transactions on starknet
+   * Contract must be deployed for fee estimation to be possible
+   *
+   * @param transactions array of transaction object containing :
+   * - type - the type of transaction : 'DECLARE' | 'DEPLOY' | 'INVOKE_FUNCTION' | 'DEPLOY_ACCOUNT'
+   * - payload - the payload of the transaction
+   *
+   * @returns response from estimate_fee
+   */
+  public abstract estimateFeeBulk(
+    transactions: TransactionBulk,
+    estimateFeeDetails?: EstimateFeeDetails
+  ): Promise<EstimateFeeResponseBulk>;
 
   /**
    * Invoke execute function in account contract
@@ -236,7 +255,7 @@ export abstract class AccountInterface extends ProviderInterface {
   - optional address salt  
   - optional contractAddress
    * @param transactionsDetail Invocation Details containing:
-  - optional nonce
+  - constant nonce = 0
   - optional version
   - optional maxFee
    * @returns a confirmation of sending a transaction on the starknet contract